Updated openstack/openstack Project: openstack/neutron  7250d99088814f653dcd75479f77a69208d7b4ae
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/559465d22e07c31d0b932bba6ba62b7b

Change-Id: I77ab202260824ec33f22a0153776fe1baa1cd120