Updated openstack/openstack Project: openstack/sahara-image-elements  66831d0d4a0e50627b89eb0d3e0944a0e558bea2
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/5a0b66960af8c723d8a2ccf0350a23d2

Change-Id: I6e991331adbf07e756bcf76606068a76129e5659