Updated openstack/openstack Project: openstack/tripleo-image-elements  b9fb5326fd4789aed2a04b30b0586aa47717d295
Copy map-services script to tripleo elements

map-services will be removed on dib v2 branch, but
tripleo still relies heavily on it. To prevent breakage,
copy the script to the internal element using it.
In the future, all elements will need to be migrated
to new svc-map, but it is a massive change.

Change-Id: I9f450de66044123ab07fa498b68f744d1029089a