Updated openstack/openstack Project: openstack-infra/project-config  a6d506f3b892edfabdb50b743d8cefdea5d8dbe9
Remove centos-7 windmill voting gate

Because we are landing zookeeper support in windmill, we don't have a
centos-7 package yet. As a result, we can not have a voting gate right
now. So we are moving centos-7 voting gates, switching them to
non-voting for now.

Additionally, promote ubuntu-trusty and ubuntu-xenial to voting
status.

Change-Id: I7501311c0b18c73f8143ef21028ae495e79cf32a
Signed-off-by: Paul Belanger <pabelanger@redhat.com>