Updated openstack/openstack Project: openstack/python-heatclient  9c62372d2dafc1d8780e00423ae64143faae8516
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/0d8a49e0bcbb30e5d69b5a89ba9db494

Change-Id: I41e74fbbd61bf46041e843ad43b5b3c4bec647b4