Updated openstack/openstack Project: openstack/tempest  ae4beb7a56bbc38a889fd55a7e017c55b82cc03d
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

b'https://gist.github.com/03d435554251cadd770793c055cf592f\n'

Change-Id: I2982952d85bf9fc2f25cb37092c8568315c24685