Updated openstack/openstack Project: openstack/keystone  641a768062a0787d47db994d333372c507a4ea54
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/fc47e904c942eace9b274fbc46f5d27d

Change-Id: I61fd31c923c100cf4c81ef2b287621193a0c6c29