Updated openstack/openstack Project: openstack/keystone  b368b42ebf4ee0dde48bb039af9453e92fd54b88
Remove eventlet-related call to sleep

Eventlet patched time.sleep. When time.sleep was called, eventlet
switched to another worker. The code being removed gave eventlet
a chance to switch to another worker if there was one.

Eventlet was removed a long time ago and the call being removed
is not needed any more.

Change-Id: I2d23f20f94c2f2e0917be9ea67e7e5edf3514aff