Updated openstack/openstack Project: openstack/nova  7b0db52b45fb06261dfe05bc043758d8373d12b9
placement: genericize on resource providers

A simple patch that makes a number of the methods of the scheduler
report client just take a resource provider UUID and a set of inventory
records. With upcoming patches, the compute node is not the only
provider of resources in the system and we want to keep various report
client methods for adding resource provider and inventory records as
generic as possible instead of being specific to a compute node.

Change-Id: Ia8fe9d71893e8bf3f4f8cddd329613e269fef210
blueprint: generic-resource-pools-ocata