Updated openstack/openstack Project: openstack/requirements  25b41410a7b46e63c4c6300ee1597ee2775e009f
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/7fa9461bc9851fdc5242f54e29ac0fb4

Change-Id: I94783fd93f760a7694760019f4a24ff122aeb5a7