Updated openstack/openstack Project: openstack/heat-translator  8e143f9f0c282e5fca46909ec2db6fc2ad267555
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/658d45aed2278ab57d76022e7d44fd15

Change-Id: I63a04e6730ee33dfa22d55580bbe20b91da3d938