Updated openstack/openstack Project: openstack/heat-templates  84d3ce7aed017cc224f9910b7b6355ff4a5c9f25
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/af4bfde1e9915d6835abcceb2907b603

Change-Id: Iafff19d2a69b5e5331632ade3071a6c8d362997c