Updated openstack/openstack Project: openstack/heat  845b82ec797e2a40aab11ce805c727350078b2b1
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/b44ce5b55c57c2fc318d27b195eb24da

Change-Id: If602066f3a499c33899deac8fc898efc34c694de