Updated openstack/openstack Project: openstack-dev/devstack  f5e78b6fb64ec45b177c2b73468f0ffe04c774e0
Allow provider network to be used for ssh validation

Currently devstack assumes that the network used for ssh
validation is the private network. This patch adds a hook that
sets the network used for ssh validation based on whether or not
provider networking is being used. It also moves the function
'is_provider_network' into functions-common as it will now be
used by both tempest and neutron.

Change-Id: I265c9e26c9bfb18b7e201f27d8912b8bec235872