Updated openstack/openstack Project: openstack/oslo.db  8be136528e99d8ea6ae8f1ae8767738b4c876759
SoftDeleteMixin: coerce deleted param to be an integer

PostgreSQL is very strict about types and won't perform an automatic
type cast when trying to pass e.g. a boolean `false` value to be
INSERT'ed or UPDATE'ed.

Coerce the value of deleted bound parameter to always be an integer
in the DB layer by adding a custom SQLAlchemy integer type decorator
to make PostgreSQL happy, should someone pass a boolean value for
deleted column (it was meant to be of boolean type, but is integer
instead to allow for Nova use case of shadow tables and compound
unique constraints on something + deleted).

Closes-Bug: #1644513

Change-Id: I13c6233cfeb611b1b106eedfc9b57d2af313c46b