Updated openstack/openstack Project: openstack/keystone  db5c3e51f81b8f65b3a9b15b23e7546e1190a9fd
Document token header in federation auth response

When reading the OS-FEDERATION API documentation, it is not clear how
to find the ID of the token when requesting either an unscoped or
scoped token. Token requests for the OS-FEDERATION API work the same
way as for the standard API, which is that the token ID is returned in
the X-Subject-Token header, so let's just mention that in the
OS-FEDERATION API documentation.

Change-Id: I6e9743d9001684f0d05ace119509f643c8b8e363