Updated openstack/openstack Project: openstack/ironic  32942aedeaebaa77feb5e56492ebedb6a8ee16d9
Drac: Deprecate drac_host property

This patch is deprecating the drac_host property in favor of
drac_address.

The drivers in the Ironic tree uses the "<namespace>_address" format for
setting the IP address to talk to their BMC except for Drac which used
"drac_host" instead. This patch is fixing it.

These small incosistencies decrease the UX and make scripting around
Ironic harder than it should be.

Change-Id: Ia23e8582a398dca9ca11762ee6fe1789fdba9777
Closes-Bug: #1644210