Updated openstack/openstack Project: openstack/horizon  6b14596a3977339559f082725e1a97c96f93304d
Trivial fixes for Swift UI

This patch adds a .row class to div#containers_wrapper that allows to reclaim
more space for the containers sidebar and the table of objects.

All other panes have their content aligned to the .page-header (h1). In this
pane (containers), its content wasn't aligned to the page-header.

Change-Id: Ie391fef808000ee2504a8570d27dec8d879bad35