Updated openstack/openstack Project: openstack/nova  64eae8dce22e4da6a824a40601849a127e1b328c
Check deleted flag in Instance.create()

An attempt to create an instance with the deleted flag set will
not work as intended, especially from instance.create(), which
can't properly pass the yet-to-be-determined deleted=id protocol.

This makes Instance.create() throw an error like the one for
being already created, and fixes the one test where we depend
on this working.

Change-Id: I60d18ca694a2eb7f333c0cf2a898cf8f17eb83dd
Closes-Bug: #1644513