Updated openstack/openstack Project: openstack/python-neutronclient  83fd3cb72c3338917d81f65462106fec937283f2
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/9c58a860190b7421066893b12675b496

Change-Id: Ie6bf3f8a7addf0b070a6447d500b5b61144c3664