Updated openstack/openstack Project: openstack-infra/devstack-gate  af04010cbad1adb87c2e1acabb85e1bb9177fe8e
Make chown and chmode to work on symlinks

When clenup functions collects logs it use recursive chown and
chmod, but it doesn't affect symlinks. There are symlinks with
absolute path that have root permissions only and are failed
to be copied with rsync later. On other hand rsync fails and
doesn't recover on this error, which fails the whole taks of
copying logs.
Let's use -L option for chown, which means traverse symlinks.
For chmod it's better to use "find" command, which will execute
on each file and directory.
See story: https://storyboard-dev.openstack.org/#!/story/55

Change-Id: I776cb8bc952c81b98cf0b1124eee9f28a92eb7b6