Updated openstack/openstack Project: openstack-infra/project-config  d900587d5fbe5a720d6e92409d067b465c90a637
Adjust filters for OpenStack-Ansible trusty/xenial jobs

Currently the filters are not working as expected. For the
openstack-ansible repository both the xenial voting and
non-voting jobs are running for master, but only the voting
one should be. For stable/newton the trusty and both xenial
voting and non-voting jobs are running, whereas it should
only be trusty and xenian-nv for the branch.

This patch re-orders the filters so that the inclusions come
first, then the skips afterwards. The inclusion of the
xenial-nv job is very specifically applied to the repository
and branch combination.

Change-Id: Ia0716bdfb5dcefa92ff4d06df8e96f19ea47c4b6