Updated openstack/openstack Project: openstack-infra/project-config  2897bea148e283a5e7cb52fec76342c109d17b8a
Add os_searchlight to OpenStack-Ansible

This patch adds new blank git repository to OpenStack-Ansible for
deploying Searchlight.

This project will have its own core team with the general
openstack-ansible core team included, with the appropriate rights to
ensure that they can act as a back-stop in case of any issues.

Needed-By: I2e5e4e7dfd12d45fe40b9dd57af56c875f2fd0d7
Change-Id: I26e29386da6661516bd61622590bdcd46d5b8102