Updated openstack/openstack Project: openstack/nova  25263890de25c68754e45e16a1d3c7a8f1bb3732
Fix expected_attrs kwarg in server_external_events

The Instance.get_by_uuid method takes an expected_attrs
kwarg which needs to be a list or tuple, not just any old
iterable like a string. Because of how the underlying
Instance object code massages this value, it's not a hard
failure but does mean you don't join the columns you expect
when getting the instance.

This makes it a list and makes sure the stub in the unit
tests is checking for valid values.

Change-Id: I3ad85f9062b5cb19962d9e6a7af52440166def45
Closes-Bug: #1645479