Updated openstack/openstack Project: openstack/nova  f47716cb80da820b8a968c8c76e2ea10242a79d4
Default deleted if the instance from BuildRequest is not having it

We should already have the rehydrated instance having the deleted field being
set. That field is also a BooleanField type, so we should never try to cast it
from an integer (even if we support it). Instead, we need to make sure that if
the field is not set for any reason (mostly because an Instance wasn't persisted
before we persisted the BuildRequest object that was nesting it), then we would
default the deleted field to False.

Change-Id: I828191f750a5978fed65a86e797388d11042b4a0
Related-Bug: #1644513