Updated openstack/openstack Project: openstack-infra/project-config  cd058b9b2e61e98d8bc6775252f22d43df9313a5
Remove special searchlight unit test setup

With the introduction of tools/test-setup.sh, we can move the setup for
searchlight to the repositories and remove the special setup for these
repos and use the normal python jobs instead.

Also, remove gate-searchlight-python27-searchlight-es1, this is broken
since August and PTL agreed that it should be removed.

Note that this is needed to fix liberty branch of searchlight - liberty
needs elasticsearch 1 while newer branches need elasticsearch 2.

Change-Id: Ie038e9c9a38727c12532bded48a76fe2cecb7fd8