Updated openstack/openstack Project: openstack/nova  997f2e0f0cec26ba47ab020818fd1c5640fe55f8
Never pass boolean deleted to instance_create()

This explicitly converts the deleted flag to an integer (which will
always be zero) at Instance.create() time. We already require it to
be False, but if set, we will pass the boolean instead of an integer,
which will confuse some (self-respecting) DBMs.

Change-Id: I67a340eea79b4423e77a00242ec6ae345e8c0e19
Closes-Bug: #1644513