Updated openstack/openstack Project: openstack/glance  c40a71e4890ad6c988cebff595cf1217c2a6264b
Prepare for oslo.log 3.17.0

The 3.17.0 release of oslo.log changes the default value of "use_stderr"
to False (from True). We had tests that asserted that our glance-*
commands used stderr to provide information to administrators. This
works to preserve that default value for operators by overriding the
default to True when we're running in one of those commands.

We deliberately excluded glance-api, glance-glare, and glance-registry
because glance-api and glance-registry are usually not run manually and
they don't need to log to stderr by default and glance-glare is
deprecated.

Change-Id: Ia054f8f637c826280722c0d2a6811fdbc0cf54ed