Updated openstack/openstack Project: openstack/trove  2171640118d7cff7f427b7947de51828e51d9c0e
Create net/subnet for alt_demo tenant

Now that devstack uses Neutron by default (and so does the gate)
we need to create a private network and subnet in the alt_demo
tenant for the int-tests to work.
(see https://review.openstack.org/#/c/355492 )
Note that we will not use this network to start as some issues
still need to be sorted out.

The tcp_ports configuration was also moved to facilitate both
network styles.

The 'PUBLIC_INTERFACE=eth0' line was also moved to be in just
nova networking as it is not necessary for Neutron (and caused
the network to become unstable).  This should allow it to
continue working for backwards compatibility if required.

A number of changes to code and/or tests were also required.
Since these are specifically tied to the switch to neutron
(as they are equally valid for nova-net) they have been
made separately and referenced below in the 'dependencies' list.
This should also facilitate easier reviews and make it
possible to backport any of the fixes if required.

The cluster tests for PXC and Redis were turned off until the
networking issues with these two datastores can be sorted out.

Merged dependencies: I136141b3db78b7a0a1804bb5e3d4ca5cf4bc3bb3
                     I47cbe9e899cb0767c377d184f123e38eda5b6a05
                     I2ede074eb9bdf26420750f19f3aa4b8d057c5d7d
                     I6aff776680c9c0632d8965c3ed70ed26db0a63f2
                     I54edc4acac09ca2278f525c08ad0d87576f0549e
                     Iab530fe503336c83e9b948c2c008d371a8030a1e
                     I6464c0031ba5b504ff103d8f62093d252deb87bb

We need to have the new variable IPV4_ADDRS_SAFE_TO_USE defined:
Depends-On: Ic6720677a4a496d246a8f9be5ede1a53821f56dc

Change-Id: I5c1714b7839b2736c50f2daa2f4506c4006815a1
Closes-Bug: #1613889