Updated openstack/openstack Project: openstack-infra/project-config  8ece36672da190df872488d2bb203c3da8df1a5b
Fix ansible skips for deploy guide

change https://review.openstack.org/404384 did not run the deploy-guide
job since there is a skip-if rule that skips the deploy-guide directory
;(

Adjust the skip-if name so that this does not happen. There is already a
specific files section for the deploy-guide to run it only if needed.

Change-Id: I32d9590dce3b258391696f7106664a23df3063d3