Updated openstack/openstack Project: openstack/trove  436eb38774b3b63465105e4e25a421d4b2887769
when pylint has an error loading the config, it should fail

When someone has a pylint config file that has an error such as when a
merge goes bad, loading the file fails, and pylint motors along
happily with the default file (i.e. nothing ignored) and this is
confusing. Unless you happen to catch the little error that it prints
out.

Instead, just bomb out with an error.

Change-Id: Ie94dda35652f2a8f06502af72b68be73d61fc94a
Closes-Bug: 1639040