Updated openstack/openstack Project: openstack/sahara-specs  fd86ea3d7bf5e928397c8b8d931c315818ae72e1
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/7ffbfb8a3a7fd95c2bfe9925cc7316e2

Change-Id: I578ab2077cc5680d9fc57e3f6f63a96e539858b7