Updated openstack/openstack Project: openstack/neutron-fwaas  ce800d817600f0364ec5231d0f7651289c4fc6a6
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/1c8b7ec0a4a119dc5e39e8a4ffde8a33

Change-Id: Ib6338d1162743dd86f6fa8b7104f5e133e4f06f8