Updated openstack/openstack Project: openstack/tripleo-puppet-elements  9ff529b832d53259e71d8931acedf64f54b63dcb
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/39dd99e3df15f4e4602c9fed19edd969

Change-Id: I272463bd05afca3b75d243e50d66d37ceac954b9