Updated openstack/openstack Project: openstack-infra/shade  1b73ba26beee187c9f8196d8958e4cb064e56381
Create and use a Adapter wrapper for REST in TaskManager

Adding a TaskManager aware wrapper around Adapter gets us a natural
interface for migrating one call at a time away from python-*client and
to direct REST calls. The Adapter shares the Session with the clients,
but it returns munches and throws exceptions that match the exceptions
we're expecting outside of shade.

Also, move compute_client calls out of their own tasks

Putting compute_client calls in a Task will cause them to be doubly
enqueued. This has to be combined with this patch because otherwise the
double-invocation fails functional tests.

Change-Id: If2d42af5fde1334b3b99ec3a9bbade38b19adbee