Updated openstack/openstack Project: openstack/tripleo-specs  2f80d6e4f245c11345adbdceeb1c4b99db1ee547
Enable Oslo.Messaging Dual Backends

This spec describes adding two functional capabilities to the messaging
services of an overcloud deployment. The first capability is to enable
the selection and configuration of separate messaging backends for
oslo.messaging RPC and Notification communications. The second
capability is to introduce support for a brokerless messaging backend
for oslo.messaging RPC communications via the AMQP 1.0 Apache
qpid-dispatch-router.

Change-Id: Idcbd264e12b28c6118cf7ff80c01a392580a58e5