Updated openstack/openstack Project: openstack-infra/shade  db84c7f0d74683674058a1f416469fad6645a7af
Have OpenStackHTTPError inherit from HTTPError

We need to have the exceptions we emit be OpenStackCloudException
exceptions, because backwards compat. However, sometimes directly
dealing with the HTTPError from requests is also a cool thing, since it
has a set of information attached to it.

Make OpenStackHTTPError multi-inherit from both of them, so that users
can catch either base exception, and also so that they can get both
behavior types.

Change-Id: I0243019991fb635302921dc1feb9be5ef6369c67