Updated openstack/openstack Project: openstack-dev/devstack  fddff69efddc1de594402a1c728b4ee40171d2b5
ceph: set rbd_secret_uuid in cinder.conf, not rbd_uuid

The ceph cinder backend script was setting the wrong
config option in cinder.conf for the secret uuid. This
was being masked by a bug in nova which is failing on
this bug when trying to fix the nova bug...right. It
makes sense.

See:

http://docs.ceph.com/docs/master/rbd/rbd-openstack/#configuring-cinder

Change-Id: I4655cae3212d589177d2570403b563a83aad529a
Closes-Bug: #1635488