Updated openstack/openstack Project: openstack/ironic  c9ddeec16d52b00d9f7206d1b1050315488238ea
Install syslinux package only for Wheezy / Trusty

Syslinux package is not available on Ubuntu Xenial ppc64el (Power). And
as of Jessie/Xenial, syslinux package only provides DOS and NTFS
loaders, which are not required. Keep installing the package for
previous releases Wheezy and Trusty to provide some compatibility
instead of removing the dependency completely.

Change-Id: I899e77423ba86cb4de2a949a29b8573d2465b76a