Updated openstack/openstack Project: openstack/python-keystoneclient  f7c1d45a04bb2b024a985b8e30245d38f2bf8442
Fix missing service_catalog parameter in Client object

Return None if service_catalog parameter does not exist.

service_catalog is the property and it is not initialized on object creation.
service_catalog tries to get value from auth_ref and raises AttributeError
exception if auth_ref is not initialized. It worked before we introduced
sessions, because authentication happened on client instantiation. Now,
when sessions are used, auth_ref is not initialized until the first
request. This change adds try-catch block in service_catalog property
to catch this error.

Change-Id: I58eb888f0989241f9e5626564bd48d901b324d36
Closes-Bug: #1508374