Updated openstack/openstack Project: openstack/swift  43a175ebd284e825c0f6e5b79a23d8a32f62326e
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/185b807657793d710f29d87ea7ec3e56

Change-Id: I78815177dba9d765587d9a0aee4e0f76e14a043e