Updated openstack/openstack Project: openstack/python-swiftclient  e3d6c34c982091882a5578ab2bdaf6da00eae4d6
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/7d9f14fa125fa35b52bc840eb94828b2

Change-Id: I5f6a3351469516aa81d5eb0fe7c842fe35d1433a