Updated openstack/openstack Project: openstack-infra/project-config  c4e7a61a89f7ca5f41a34c9d4b7a7bc4a3012bc4
Disable devstack plugin for builder 'ironic-grenade'

Disable the devstack plugin for items which use the 'ironic-grenade'
builder as the grenade plugin in Ironic already enables the devstack
plugin. If the devstack plugin is enabled twice it gets run twice
which will have it attempting to create the bare-metal VMs twice and
failing.

Create a new builder: 'ironic-vars-early' as the ordering is
important.  Previously we had a bug because ordering was changed and
the ironic plugin was enabled after the ironic-inspector plugin. Work
around this issue by using this new ironic-vars-early builder.

Change-Id: I6b4c2b1c3149a185d0495ed6f7674e173614f63b