Updated openstack/openstack Project: openstack/cinder  366acd02b40f5b25bcc8cbbd9a89623feb7fc884
Support A/A in attach/detach operations

This patch allows attaching and detaching volumes in HA Active-Active
configurations.

To allow the attach and detach of volumes in Cinder using the cluster
topic queue in the Message Broker following RPC methods in
cinder/volume/rpcapi has been changed:

- attach_volume
- detach_volume
- initialize_connection
- terminate_connection
- remove_export

When the cluster field is not set for a volume all RPC calls will behave
as before, sending it to the host topic queue.

Specs: https://review.openstack.org/327283
Implements: blueprint cinder-volume-active-active-support
Change-Id: I6ec689ea8abb10832a47b20f82bf2e710c62deae