Updated openstack/openstack Project: openstack/heat  17ec3a3b959385500e4fc704449b8b394005f9a1
Make the name of a Zaqar queue optional

There's no reason to force the user to specify the name of a Zaqar queue.
In fact, doing so makes it difficult for users to create multiple instance
of the same template without having all of the messages go into a single
queue (and without deleting the shared queue when one of the stacks is
deleted).

We now autogenerate a sensible randomised name if none is provided.

Change-Id: I5195d8c0dd413a8f1bd28c2a8ef7a3191c74f882
Partial-Bug: #1637304