Updated openstack/openstack Project: openstack-infra/tripleo-ci  a786b9083df9ec3201ea931b723dd809124735d9
Use command instead of env var for setting the hostname in the undercloud

the $HOSTNAME environment variable gets the hostname from at bootup
and doesn't take into account if the hostname was changed. Using the
hostname command takes this into account.

So, this effectively adds the node's FQDN as the canonical hostname,
leaving the rest of the entries untouched, so they would still be
accesible, but as aliases.

This also takes IPv6 into account.

Change-Id: I4bf8dbd02bdf6a283f57b930a8d43c6ebc6253b8