Updated openstack/openstack Project: openstack/neutron  5fbe5212ce5f4d3ba8ae3d8e74f2b0f6509bd82d
Functional tests: change assert to wait_until_true

In certain cases, the keepalived can take a few seconds to run/stop.
For this reason, the patch proposes a change to the tests, in which
instead of asserting that the process started/stopped, we'll wait a
short while until they do. This will cause failing tests to not fail
simply because it took a few extra CPU cycles to run.

Change-Id: I497ef32651fbb68bf49e172543f9675f6279df4b