Updated openstack/openstack Project: openstack/ironic  f857a883d38ad152bb6ae1739127672173b2a564
Make all IronicExceptions RPC-serializable

Currently, if a random object is passed to the IronicException
constructor, and then such exception needs to be sent over RPC,
its serialization fails. In case of this message was supposed
to be sent from conductor to API as response to an RPC call,
API times out waiting for response and throws HTTP 500 error.
This change makes sure that everything passed to the exception's
constructor in kwargs is serializable, otherwise it's dropped
from kwargs.

Co-Authored-By: Richard Pioso <richard.pioso@dell.com>
Closes-Bug: #1639982
Change-Id: Icc153c6eefec7bd1451bc2110f742ae0241d49a1